requirements.txt
slackclient
discord.py
regex